rest/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
torch==2.9.0
torchvision==0.24.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
